duckdb-vs-polars-vs-fireducks/pyproject.toml
[project]
name = "duckdb-vs-polars-vs-fireducks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "contourpy>=1.3.1",
    "cycler>=0.12.1",
    "duckdb>=1.2.0",
    "fireducks>=1.2.2",
    "fonttools>=4.56.0",
    "iniconfig>=2.0.0",
    "kiwisolver>=1.4.8",
    "matplotlib>=3.10.0",
    "numpy>=2.2.3",
    "packaging>=24.2",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pluggy>=1.5.0",
    "polars-lts-cpu>=1.22.0",
    "pyarrow>=19.0.0",
    "pyparsing>=3.2.1",
    "pytest>=8.3.4",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2025.1",
    "seaborn>=0.13.2",
    "six>=1.17.0",
    "tzdata>=2025.1",
]
